+++ src/options.cpp
@@ -1066,21 +1066,6 @@ std::vector<options_manager::id_and_option> options_manager::build_soundpacks_li
     return result;
 }
 
-#if defined(__ANDROID__)
-bool android_get_default_setting( const char *settings_name, bool default_value )
-{
-    JNIEnv *env = ( JNIEnv * )SDL_AndroidGetJNIEnv();
-    jobject activity = ( jobject )SDL_AndroidGetActivity();
-    jclass clazz( env->GetObjectClass( activity ) );
-    jmethodID method_id = env->GetMethodID( clazz, "getDefaultSetting", "(Ljava/lang/String;Z)Z" );
-    jboolean ans = env->CallBooleanMethod( activity, method_id, env->NewStringUTF( settings_name ),
-                                           default_value );
-    env->DeleteLocalRef( activity );
-    env->DeleteLocalRef( clazz );
-    return ans;
-}
-#endif
-
 void options_manager::Page::removeRepeatedEmptyLines()
 {
     const auto empty = [&]( const PageItem & it ) -> bool {
