+++ src/options_android.cpp
@@ -1,6 +1,26 @@
 #include "options.h"
 #include "options_category.h"
 
+#if defined(__ANDROID__)
+#   include "options_android.h"
+#endif
+
+#if defined(__ANDROID__)
+bool android_get_default_setting( const char *settings_name, bool default_value )
+{
+    JNIEnv *env = ( JNIEnv * )SDL_AndroidGetJNIEnv();
+    jobject activity = ( jobject )SDL_AndroidGetActivity();
+    jclass clazz( env->GetObjectClass( activity ) );
+    jmethodID method_id = env->GetMethodID( clazz, "getDefaultSetting", "(Ljava/lang/String;Z)Z" );
+    jboolean ans = env->CallBooleanMethod( activity, method_id, env->NewStringUTF( settings_name ),
+                                           default_value );
+    env->DeleteLocalRef( activity );
+    env->DeleteLocalRef( clazz );
+    return ans;
+}
+#endif
+
+
 auto options_manager::add_options_android() -> void
 {
 #if defined(__ANDROID__)
