+++ src/options_android.h
@@ -1,0 +1,7 @@
+#pragma once
+#ifndef CATA_SRC_OPTIONS_ANDROID_H
+#define CATA_SRC_OPTIONS_ANDROID_H
+
+bool android_get_default_setting( const char *settings_name, bool default_value );
+
+#endif // CATA_SRC_OPTIONS_ANDROID_H
